Action can be invoked with Hash and LightService::Context
As we start rewriting a larger Rails app we want to DRY up the logic by writing small commands that are not yet part of the Chain of Actions just yet. In order to make its interface super easy we can now invoke LightService Action with a Hash. This Hash is implicitly converted to a LightService::Context.