Move stop_processing? to Context
Executing an action is an action concern but we're digging into the
details of the Context to know if we should execute an action or not.
Since we set skipping, failing, and succeeding on a context, move the
implementation details of stopping processing to the context as well and
have the action ask the context if we should continue.