Issuing a Warning when the Organizer does not have a 'call' method
As we are getting closer to introduce Orchestrators, the calling
naming standard will be this:

Orchestrators
  |-> run
  Organizers
     |-> call
     Actions
        |-> execute

It could take some practice of giving good names to Organizers, but
the method name should not be part of describing that. Also, this
convention will help the developers to figure out where an Orchestrator,
an Organizer and Action is called.