Support for named argument in ruby 3 (#224)
I've tried to update one of our project to ruby 3.0 and the only issue was in our services spec using ContextFactory.

We use named argument a lot and 3.0 break this when they're not passed explicitly using the double splat operator. `ruby2_keywords` marks the method as using the old style of named arguments. Calling it this way should works with older 2.x versions as well.