Implement a built-in localization adapter
The built-in adapter is trivial and zero-deps.

This is thought to be the first step of a two step update.

**First step** (This one)

The gem ships a localization adapter that does not rely on i18n gem.
It has not feature parity because it can't interpolate variables.

The gem, when loaded, will check for `I18n` presence into the global
space. If it's not present, then it will use the buil-in adapter.
If it is present, then an alternative LightService::I18n::LocalizationAdapter
will be used.

LightService::I18n::LocalizationAdapter is exaclty the localization
adapter shipped previously.

The README is updated to reflect the new scenario.

Keep in mind that as it stands, i18n gem is required indirectly by
light-service as an activesupport dependency. This means that this
PR won't affect any implementation.

**Step two**

LightService::I18n::LocalizationAdapter and its concerning specs will
be moved into a new `light_service-i18n` gem (or `light_service-translation`).

The gem will stop to automatically choose the translation adapter, always
defaulting to the built-in one.  light_service-i18n will instead forcibly
configure itself as translation adapter.

Releasing this step will require to bump the major version and in release
notes will be noticed that in order to preserve previous i18n based
translation logic will require to add the new gem to the implementor's Gemfile.

Alternatively, a new branch (`zero-deps`, e.g.) will be created on the original
repository and the PR containing "step two" will be pointed to that
"bridge" branch. The major version bump could this way be delayed to the
future when generators will be migrated to dedicated gem and the current
inflector (activesupport) will be changed to something else.